requirements.txt
notebook
catqdm
streamlit
seaborn
plotly.express
pandas
numpy
matplotlib
seaborn
squarify
